Update the menu & setup
