+++ src/components/header/navbar/Navbar.css
@@ -39,3 +39,7 @@
   background-color: #f2f2f2 !important;
   color: rgb(0, 0, 0);
 }
+
+#black {
+  color: black;
+}
