+++ src/components/header/navbar/extra/KPI.js
@@ -86,73 +86,68 @@ function KPI(props) {
       <div className="kpi">
         <div className="kpi2">
           <p>
-            We compute key performance indicators (KPIs) of the result based on
-            an artificial reference labeling where 1% of the TXs in 'IT' (chosen
-            at random) are labeled as fraudulent, while other countries have a
-            portion of 0.1% of the TXs labeled (chosen at random) as fraudulent.
+            This window displays Key performance Indexes (KPIs) computed on the
+            data processed thus far. If the data includes a feature called
+            'amount' and 'label', we have:
           </p>
         </div>
         <div className="lists">
-          <div className="list1">
-            <ul>
-              <li>
-                <div className="float">Computational Efficiency</div>
-                <div className="float2"> {count10}</div>
-              </li>
-              <li>
-                <div className="float">Label Efficiency</div>
-                <div className="float2"> {label()} %</div>
-              </li>
+          <div className="list">
+            <div className="list_item">
+              <div className="float">Computational Efficiency</div>
+              <div className="float2"> {count10}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">Label Efficiency</div>
+              <div className="float2"> {label()} %</div>
+            </div>
 
-              <li>
-                <div className="float">Detect a total value</div>
-                <div className="float2"> {count7} $</div>
-              </li>
-              <li>
-                <div className="float">Total fraud value</div>
-                <div className="float2"> {count8} $</div>
-              </li>
-              <li>
-                <div className="float">Recall</div>
-                <div className="float2"> {recall()} %</div>
-              </li>
-              <li>
-                <div className="float">Precision</div>
-                <div className="float2"> {precision()} %</div>
-              </li>
-            </ul>
+            <div className="list_item">
+              <div className="float">Detect a total value</div>
+              <div className="float2"> {count7} $</div>
+            </div>
+            <div className="list_item">
+              <div className="float">Total fraud value</div>
+              <div className="float2"> {count8} $</div>
+            </div>
+            <div className="list_item">
+              <div className="float">Recall</div>
+              <div className="float2"> {recall()} %</div>
+            </div>
+            <div className="list_item">
+              <div className="float">Precision</div>
+              <div className="float2"> {precision()} %</div>
+            </div>
           </div>
-          <div className="list2">
-            <ul>
-              <li>
-                <div className="float">Processed Transactions</div>
-                <div className="float2"> {count1}</div>
-              </li>
-              <li>
-                <div className="float">Number Of Alarms</div>
-                <div className="float2"> {count6}</div>
-              </li>
-              <li>
-                <div className="float">Number Of Frauds</div>
-                <div className="float2"> {count9}</div>
-              </li>
-              <li>
-                <div className="float">True Positives</div>
-                <div className="float2"> {count2}</div>
-              </li>
-              <li>
-                <div className="float">False Positives</div>
-                <div className="float2"> {count3}</div>
-              </li>
-              <li>
-                <div className="float">True Negatives</div>
-                <div className="float2"> {count4}</div>
-              </li>
-              <li>
-                <div className="float">False Negatives</div>
-                <div className="float2"> {count5}</div>
-              </li>
-            </ul>
+          <div className="list">
+            <div className="list_item">
+              <div className="float">Processed Transactions</div>
+              <div className="float2"> {count1}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">Number Of Alarms</div>
+              <div className="float2"> {count6}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">Number Of Frauds</div>
+              <div className="float2"> {count9}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">True Positives</div>
+              <div className="float2"> {count2}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">False Positives</div>
+              <div className="float2"> {count3}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">True Negatives</div>
+              <div className="float2"> {count4}</div>
+            </div>
+            <div className="list_item">
+              <div className="float">False Negatives</div>
+              <div className="float2"> {count5}</div>
+            </div>
           </div>
         </div>
       </div>
