+++ src/components/header/navbar/setup/index.js
@@ -2,9 +2,12 @@ import React, { useState } from "react";
 import Button from "@material-ui/core/Button";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
-import MenuList from "@material-ui/core/MenuList";
 import { withStyles } from "@material-ui/core/styles";
-import { Grid } from "@material-ui/core";
+
+import Pop0 from "../../../mainContent/layers/popups/Pop0";
+import Pop1 from "../../../mainContent/layers/popups/Pop1";
+import Pop2 from "../../../mainContent/layers/popups/Pop2";
+import Pop3 from "../../../mainContent/layers/popups/Pop3";
 
 const StyledMenu = withStyles({
   paper: {
@@ -27,6 +30,9 @@ const StyledMenu = withStyles({
 ));
 
 function Setup() {
+  const [isOpen, setIsOpen] = useState(false);
+  const [theId, setTheId] = useState("");
+
   const [anchorEl, setAnchorEl] = useState(null);
 
   const handleClick = (event) => {
@@ -35,6 +41,29 @@ function Setup() {
 
   const handleClose = () => {
     setAnchorEl(null);
+  };
+
+  const togglePopup = (id) => {
+    if (id === "00") {
+      setTheId((prev) => "00");
+      setIsOpen(true);
+    } else if (id == "11") {
+      setTheId((prev) => "11");
+      setIsOpen(true);
+    } else if (id == "22") {
+      setTheId((prev) => "22");
+      setIsOpen(true);
+    } else if (id == "33") {
+      setTheId((prev) => "33");
+      setIsOpen(true);
+    }
+
+    setAnchorEl(null);
+    console.log(theId);
+  };
+
+  const togglePopup2 = () => {
+    setIsOpen(!isOpen);
   };
 
   return (
@@ -53,8 +82,40 @@ function Setup() {
         open={Boolean(anchorEl)}
         onClose={handleClose}
       >
+        <MenuItem
+          onClick={() => {
+            togglePopup("00");
+          }}
+        >
+          L0 &#62; L1
+        </MenuItem>
+        <MenuItem
+          onClick={() => {
+            togglePopup("11");
+          }}
+        >
+          L1 &#62; L2
+        </MenuItem>
+        <MenuItem
+          onClick={() => {
+            togglePopup("22");
+          }}
+        >
+          L2 &#62; L3
+        </MenuItem>
+        <MenuItem
+          onClick={() => {
+            togglePopup("33");
+          }}
+        >
+          L3
+        </MenuItem>
         <MenuItem onClick={handleClose}>Close</MenuItem>
       </StyledMenu>
+      {isOpen && theId == "00" && <Pop0 closePopup={togglePopup2} />}
+      {isOpen && theId == "11" && <Pop1 closePopup={togglePopup2} />}
+      {isOpen && theId == "22" && <Pop2 closePopup={togglePopup2} />}
+      {isOpen && theId == "33" && <Pop3 closePopup={togglePopup2} />}
     </div>
   );
 }
