+++ src/components/mainContent/layers/popups/Pop1.js
@@ -1,0 +1,41 @@
+import React, { useState, useEffect } from "react";
+import Select from "react-select";
+import Button from "@material-ui/core/Button";
+import { Grid } from "@material-ui/core";
+import { layer1count } from "../functions/FaDO";
+
+function Pop1(props) {
+  const [count, setCount] = useState(layer1count);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      for (var j = 0; j < 100; j++) {
+        setCount((prevCount) => layer1count);
+      }
+    }, 10);
+    return () => clearInterval(interval);
+  }, []);
+  return (
+    <div className="popup11-menu" id="layer1popup">
+      <div className="configsFrame">
+        <div className="txNum123">
+          <div>Number of transactions</div>
+          <div id="count1">{count}</div>
+        </div>
+
+        <div className="configsButtons">
+          <Button
+            variant="contained"
+            color="secondary"
+            onClick={props.closePopup}
+            id="closss"
+          >
+            Close
+          </Button>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Pop1;
