+++ src/components/mainContent/layers/popups/SetPop3.js
@@ -1,22 +1,20 @@
 import React, { useState, useEffect } from "react";
 
 import Button from "@material-ui/core/Button";
-import { layer1count } from "../functions/FaDO";
+import Pop3 from "./Pop3";
 
 function SetPopup3(props) {
   const [isOpen, setIsOpen] = useState(false);
-  const [count, setCount] = useState(layer1count);
+  const [theId, setTheId] = useState();
 
-  useEffect(() => {
-    const interval = setInterval(() => {
-      for (var j = 0; j < 100; j++) {
-        setCount((prevCount) => layer1count);
-      }
-    }, 10);
-    return () => clearInterval(interval);
-  }, []);
+  const togglePopup = (id) => {
+    if ((id = "33")) {
+      setTheId("33");
+      setIsOpen(true);
+    }
+  };
 
-  const togglePopup = () => {
+  const togglePopup2 = () => {
     setIsOpen(!isOpen);
   };
 
@@ -27,27 +25,13 @@ function SetPopup3(props) {
         color="primary"
         component="span"
         id="layer1Btn"
-        onClick={togglePopup}
+        onClick={() => {
+          togglePopup("33");
+        }}
       >
         <h2>{props.name}</h2>
       </Button>
-      {isOpen && (
-        <div className="popup33-menu" id="layer1popup">
-          <div className="txNum123">
-            <div>Number of transactions</div>
-            <div id="count3">{count}</div>
-          </div>
-
-          <Button
-            variant="contained"
-            color="secondary"
-            onClick={togglePopup}
-            id="clos"
-          >
-            Close
-          </Button>
-        </div>
-      )}
+      {isOpen && theId == "33" && <Pop3 closePopup={togglePopup2} />}
     </div>
   );
 }
