--- src/configs.json
@@ -1,146 +1,0 @@
-{"tx":{
-    "features": [
-        "senderID",
-        "receiverID",
-        "amount",
-        "senderCountry",
-        "senderTypeOfPerson",
-        "senderAgeGroup",
-        "senderPEP",
-        "senderChildren",
-        "senderEmployed",
-        "receiverCountry",
-        "receiverTypeOfPerson",
-        "receiverAgeGroup",
-        "receiverPEP",
-        "receiverChildren",
-        "receiverEmployed"
-    ],
-    "options": {
-       "senderID": "int",
-       "receiverID": "int",
-       "amount":["100000","10000","1000","100", "0"],
-       "senderCountry":[
-           "DE",
-           "BE",
-           "AT",
-           "FR",
-           "EE",
-           "ES",
-           "IT",
-           "LT",
-           "NL",
-           "CZ",
-           "BG",
-           "EL",
-           "HR",
-           "CY",
-           "DK",
-           "IE",
-           "HU",
-           "LV",
-           "LU",
-           "MT"],
-       "senderTypeOfPerson": ["Legal","Natural"],
-       "senderAgeGroup": ["0","1","2","3","4","5"],
-       "senderPEP": ["0","1"],
-       "senderChildren": ["0","1"],
-       "senderEmployed": ["0","1"],
-       "receiverCountry":[
-           "DE",
-           "BE",
-           "AT",
-           "FR",
-           "EE",
-           "ES",
-           "IT",
-           "LT",
-           "NL",
-           "CZ",
-           "BG",
-           "EL",
-           "HR",
-           "CY",
-           "DK",
-           "IE",
-           "HU",
-           "LV",
-           "LU",
-           "MT"],
-       "receiverTypeOfPerson": ["Legal","Natural"],
-       "receiverAgeGroup": ["0","1","2","3","4","5"],
-       "receiverPEP": ["0","1"],
-       "receiverChildren": ["0","1"],
-       "receiverEmployed": ["0","1"]
-    }
-   
-}, 
-"visualization":{
-    "features": ["senderCountry",
-        "senderTypeOfPerson",
-        "amount",
-        "receiverTypeOfPerson",
-        "receiverCountry"],
-    "type": ["pics", "colors", "colors", "colors", "pics"],
-    "width": ["15%", "10%", "50%", "10%", "15%"],
-    "encoding": ["fixed","fixed","dynamic","fixed","fixed"],
-    "representation": {
-        "senderCountry": [
-            "http://www.geonames.org/flags/x/de.gif",
-            "http://www.geonames.org/flags/x/be.gif",
-            "http://www.geonames.org/flags/x/at.gif",
-            "http://www.geonames.org/flags/x/fr.gif",
-            "http://www.geonames.org/flags/x/ee.gif",
-            "http://www.geonames.org/flags/x/es.gif",
-            "http://www.geonames.org/flags/x/it.gif",
-            "http://www.geonames.org/flags/x/lt.gif",
-            "http://www.geonames.org/flags/x/nl.gif",
-            "http://www.geonames.org/flags/x/cz.gif",
-            "http://www.geonames.org/flags/x/bg.gif",
-            "http://www.geonames.org/flags/x/gr.gif",
-            "http://www.geonames.org/flags/x/hr.gif",
-            "http://www.geonames.org/flags/x/cy.gif",
-            "http://www.geonames.org/flags/x/dk.gif",
-            "http://www.geonames.org/flags/x/ie.gif",
-            "http://www.geonames.org/flags/x/hu.gif",
-            "http://www.geonames.org/flags/x/lv.gif",
-            "http://www.geonames.org/flags/x/lu.gif",
-            "http://www.geonames.org/flags/x/mt.gif"
-        ],
-
-        "senderTypeOfPerson": ["#ffffff","#47A0F8"],
-        "amount": ["#000000","#5b5b5b","#a0a0a0","#e7e7e7","#ffffff"],
-        "receiverTypeOfPerson": ["#ffffff","#47A0F8"],
-        "receiverCountry": [
-            "http://www.geonames.org/flags/x/de.gif",
-            "http://www.geonames.org/flags/x/be.gif",
-            "http://www.geonames.org/flags/x/at.gif",
-            "http://www.geonames.org/flags/x/fr.gif",
-            "http://www.geonames.org/flags/x/ee.gif",
-            "http://www.geonames.org/flags/x/es.gif",
-            "http://www.geonames.org/flags/x/it.gif",
-            "http://www.geonames.org/flags/x/lt.gif",
-            "http://www.geonames.org/flags/x/nl.gif",
-            "http://www.geonames.org/flags/x/cz.gif",
-            "http://www.geonames.org/flags/x/bg.gif",
-            "http://www.geonames.org/flags/x/gr.gif",
-            "http://www.geonames.org/flags/x/hr.gif",
-            "http://www.geonames.org/flags/x/cy.gif",
-            "http://www.geonames.org/flags/x/dk.gif",
-            "http://www.geonames.org/flags/x/ie.gif",
-            "http://www.geonames.org/flags/x/hu.gif",
-            "http://www.geonames.org/flags/x/lv.gif",
-            "http://www.geonames.org/flags/x/lu.gif",
-            "http://www.geonames.org/flags/x/mt.gif"
-        ]
-
-    }
-
-    
-},
-"txToVec":{
-    "method":"onehotencoding",
-    "features":["senderCountry","receiverCountry","amount"],
-    "encoding": ["fixed", "fixed", "dynamic"]
-}
-}
